+++ packages/api-client/home/index.ts
@@ -1,1 +1,0 @@
-export { getHomeStatistics } from "./get-home-statistics";